Organizando Tipagens após Reunião

+++ src/main/java/com/upe/brinquedotecaapi/model/Child.java
@@ -10,24 +10,24 @@ public class Child {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private  int id;
+    private Long id;
     private String observations;
     private String description;
 
     public Child() {
     }
 
-    public Child(int id, String observations, String description) {
+    public Child(Long id, String observations, String description) {
         this.id = id;
         this.observations = observations;
         this.description = description;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
